A sample dataset (Question 2)
This is a dataset that can be used to test the working of the solution to problem 2. The program would work as intended using any other dataset as well, only provided that the data is input in the same format as mentioned here.
This sample includes a collection of 50 Hindi songs, along with their associated artists, album, genre and duration.